# lib/api-zod/src/generated/types/backup.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Web Hosting Control Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface Backup {
  id: number;
  name: string;
  type: string;
  size: number;
  status: string;
  createdAt: string;
}