# lib/api-zod/src/generated/types/domain.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Web Hosting Control Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface Domain {
  id: number;
  name: string;
  documentRoot: string;
  phpVersion: string;
  status: string;
  sslEnabled?: boolean;
  createdAt: string;
}